pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel", "setuptools_scm>=6.4"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "src/tswf/_version.py"

[project]
name = "tswf"
description = "Tree sequence workflow"
readme = "README.md"
keywords = ["bioinformatics"]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Bio-Informatics"
]
dependencies = [
    'click>=8.0',
    'importlib-metadata; python_version>="3.7"',
    'jinja2',
    'jsonschema',
    'pypandoc',
    'ruamel.yaml'
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = {attr = "tswf.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-data = { tswf = [
	     "workflow/Snakefile",
	     "workflow/envs/*",
	     "workflow/rules/*",
	     "workflow/schemas/*",
	     "templates/*.j2",
	     "config/*",
	     "resources/*",
	     "test_wf/*"
]}


[project.scripts]
tswf = "tswf.cli:cli"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --doctest-modules"
pythonpath = [
    "src"
]
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]


[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py39
isolated_build = true

[testenv]
changedir = tests
deps =
    pytest
    pytest-cov
    pip>=7.0.0
    wheel
commands =
    pytest --basetemp="{envtmpdir}" {posargs}
"""

[tool.snakefmt]
line_length = 90
include = '\.smk$|^Snakefile|\.py$'

# snakefmt passes these options on to black
[tool.black]
skip_string_normalization = true
